fast/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.7
exceptiongroup==1.2.0
fastapi==0.104.1
fonttools==4.47.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
importlib-resources==6.1.1
joblib==1.3.2
jose==1.0.0
kiwisolver==1.4.5
matplotlib==3.8.2
mysql-connector-python==8.2.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
patsy==0.5.4
Pillow==10.1.0
pmdarima==2.0.4
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
statsmodels==0.14.1
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
zipp==3.17.0
